+++ config.w32
@@ -394,8 +394,6 @@ if (PHP_GRPC != "no") {
     "src\\core\\ext\\xds\\xds_api.cc " +
     "src\\core\\ext\\xds\\xds_bootstrap.cc " +
     "src\\core\\ext\\xds\\xds_certificate_provider.cc " +
-    "src\\core\\ext\\xds\\xds_channel_creds.cc " +
-    "src\\core\\ext\\xds\\xds_channel_default_creds.cc " +
     "src\\core\\ext\\xds\\xds_channel_stack_modifier.cc " +
     "src\\core\\ext\\xds\\xds_client.cc " +
     "src\\core\\ext\\xds\\xds_client_stats.cc " +
@@ -633,6 +631,8 @@ if (PHP_GRPC != "no") {
     "src\\core\\lib\\security\\credentials\\tls\\grpc_tls_credentials_options.cc " +
     "src\\core\\lib\\security\\credentials\\tls\\tls_credentials.cc " +
     "src\\core\\lib\\security\\credentials\\tls\\tls_utils.cc " +
+    "src\\core\\lib\\security\\credentials\\xds\\xds_channel_creds.cc " +
+    "src\\core\\lib\\security\\credentials\\xds\\xds_channel_default_creds.cc " +
     "src\\core\\lib\\security\\credentials\\xds\\xds_credentials.cc " +
     "src\\core\\lib\\security\\security_connector\\alts\\alts_security_connector.cc " +
     "src\\core\\lib\\security\\security_connector\\fake\\fake_security_connector.cc " +
