+++ gRPC-Core.podspec
@@ -944,9 +944,6 @@ Pod::Spec.new do |s|
                       'src/core/ext/xds/xds_certificate_provider.cc',
                       'src/core/ext/xds/xds_certificate_provider.h',
                       'src/core/ext/xds/xds_channel_args.h',
-                      'src/core/ext/xds/xds_channel_creds.cc',
-                      'src/core/ext/xds/xds_channel_creds.h',
-                      'src/core/ext/xds/xds_channel_default_creds.cc',
                       'src/core/ext/xds/xds_channel_stack_modifier.cc',
                       'src/core/ext/xds/xds_channel_stack_modifier.h',
                       'src/core/ext/xds/xds_client.cc',
@@ -1409,6 +1406,9 @@ Pod::Spec.new do |s|
                       'src/core/lib/security/credentials/tls/tls_credentials.h',
                       'src/core/lib/security/credentials/tls/tls_utils.cc',
                       'src/core/lib/security/credentials/tls/tls_utils.h',
+                      'src/core/lib/security/credentials/xds/xds_channel_creds.cc',
+                      'src/core/lib/security/credentials/xds/xds_channel_creds.h',
+                      'src/core/lib/security/credentials/xds/xds_channel_default_creds.cc',
                       'src/core/lib/security/credentials/xds/xds_credentials.cc',
                       'src/core/lib/security/credentials/xds/xds_credentials.h',
                       'src/core/lib/security/security_connector/alts/alts_security_connector.cc',
@@ -2022,7 +2022,6 @@ Pod::Spec.new do |s|
                               'src/core/ext/xds/xds_bootstrap.h',
                               'src/core/ext/xds/xds_certificate_provider.h',
                               'src/core/ext/xds/xds_channel_args.h',
-                              'src/core/ext/xds/xds_channel_creds.h',
                               'src/core/ext/xds/xds_channel_stack_modifier.h',
                               'src/core/ext/xds/xds_client.h',
                               'src/core/ext/xds/xds_client_stats.h',
@@ -2248,6 +2247,7 @@ Pod::Spec.new do |s|
                               'src/core/lib/security/credentials/tls/grpc_tls_credentials_options.h',
                               'src/core/lib/security/credentials/tls/tls_credentials.h',
                               'src/core/lib/security/credentials/tls/tls_utils.h',
+                              'src/core/lib/security/credentials/xds/xds_channel_creds.h',
                               'src/core/lib/security/credentials/xds/xds_credentials.h',
                               'src/core/lib/security/security_connector/alts/alts_security_connector.h',
                               'src/core/lib/security/security_connector/fake/fake_security_connector.h',
