+++ grpc.gemspec
@@ -863,9 +863,6 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/xds/xds_certificate_provider.cc )
   s.files += %w( src/core/ext/xds/xds_certificate_provider.h )
   s.files += %w( src/core/ext/xds/xds_channel_args.h )
-  s.files += %w( src/core/ext/xds/xds_channel_creds.cc )
-  s.files += %w( src/core/ext/xds/xds_channel_creds.h )
-  s.files += %w( src/core/ext/xds/xds_channel_default_creds.cc )
   s.files += %w( src/core/ext/xds/xds_channel_stack_modifier.cc )
   s.files += %w( src/core/ext/xds/xds_channel_stack_modifier.h )
   s.files += %w( src/core/ext/xds/xds_client.cc )
@@ -1328,6 +1325,9 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/security/credentials/tls/tls_credentials.h )
   s.files += %w( src/core/lib/security/credentials/tls/tls_utils.cc )
   s.files += %w( src/core/lib/security/credentials/tls/tls_utils.h )
+  s.files += %w( src/core/lib/security/credentials/xds/xds_channel_creds.cc )
+  s.files += %w( src/core/lib/security/credentials/xds/xds_channel_creds.h )
+  s.files += %w( src/core/lib/security/credentials/xds/xds_channel_default_creds.cc )
   s.files += %w( src/core/lib/security/credentials/xds/xds_credentials.cc )
   s.files += %w( src/core/lib/security/credentials/xds/xds_credentials.h )
   s.files += %w( src/core/lib/security/security_connector/alts/alts_security_connector.cc )
