+++ grpc.gyp
@@ -860,8 +860,6 @@
         'src/core/ext/xds/xds_api.cc',
         'src/core/ext/xds/xds_bootstrap.cc',
         'src/core/ext/xds/xds_certificate_provider.cc',
-        'src/core/ext/xds/xds_channel_creds.cc',
-        'src/core/ext/xds/xds_channel_default_creds.cc',
         'src/core/ext/xds/xds_channel_stack_modifier.cc',
         'src/core/ext/xds/xds_client.cc',
         'src/core/ext/xds/xds_client_stats.cc',
@@ -1055,6 +1053,8 @@
         'src/core/lib/security/credentials/tls/grpc_tls_credentials_options.cc',
         'src/core/lib/security/credentials/tls/tls_credentials.cc',
         'src/core/lib/security/credentials/tls/tls_utils.cc',
+        'src/core/lib/security/credentials/xds/xds_channel_creds.cc',
+        'src/core/lib/security/credentials/xds/xds_channel_default_creds.cc',
         'src/core/lib/security/credentials/xds/xds_credentials.cc',
         'src/core/lib/security/security_connector/alts/alts_security_connector.cc',
         'src/core/lib/security/security_connector/fake/fake_security_connector.cc',
@@ -1347,7 +1347,6 @@
         'src/core/ext/upb-generated/src/proto/grpc/lb/v1/load_balancer.upb.c',
         'src/core/ext/upb-generated/validate/validate.upb.c',
         'src/core/ext/upb-generated/xds/data/orca/v3/orca_load_report.upb.c',
-        'src/core/ext/xds/xds_channel_creds.cc',
         'src/core/lib/address_utils/parse_address.cc',
         'src/core/lib/address_utils/sockaddr_utils.cc',
         'src/core/lib/backoff/backoff.cc',
@@ -1490,6 +1489,7 @@
         'src/core/lib/security/credentials/composite/composite_credentials.cc',
         'src/core/lib/security/credentials/credentials.cc',
         'src/core/lib/security/credentials/plugin/plugin_credentials.cc',
+        'src/core/lib/security/credentials/xds/xds_channel_creds.cc',
         'src/core/lib/security/security_connector/security_connector.cc',
         'src/core/lib/security/transport/client_auth_filter.cc',
         'src/core/lib/security/transport/secure_endpoint.cc',
