+++ src/core/lib/config/core_configuration.h
@@ -19,9 +19,9 @@
 
 #include <atomic>
 
-#include "src/core/ext/xds/xds_channel_creds.h"
 #include "src/core/lib/channel/channel_args_preconditioning.h"
 #include "src/core/lib/channel/handshaker_registry.h"
+#include "src/core/lib/security/credentials/xds/xds_channel_creds.h"
 #include "src/core/lib/surface/channel_init.h"
 
 namespace grpc_core {
